Fix jquery RadioGropu component
