+++ packages/jquery/src/radio-button.js
@@ -3,20 +3,6 @@ import { ComponentWrapper } from './component-wrapper';
 import { RadioButton as RadioButtonInferno } from './generated/components/radio-button';
 
 export class RadioButton extends ComponentWrapper {
-  _initializeComponent() {
-    super._initializeComponent();
-    this._propsInfo.templates.forEach((template) => {
-      this._componentTemplates[template] = this._createTemplateComponent(
-        this._props[template],
-        template,
-      );
-    });
-  }
-
-  getProps() {
-    return super.getProps();
-  }
-
   get _propsInfo() {
     return {
       twoWay: [['checked', 'defaultChecked']],
