+++ packages/react/src/components/radio-group/index.ts
@@ -1,1 +1,4 @@
+import type { RadioGroupProps, RadioGroupRef } from './types';
+
 export * from './radio-group';
+export { RadioGroupProps, RadioGroupRef };
